refactor: 멤버의 첫 달력 정보 불러오기 로직 수행 위치 변경
